requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
deepl==1.21.0
greenlet==3.1.1
idna==3.10
marshmallow==3.26.1
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
passlib[bcrypt]==1.7.4